Home page is now ssr
